fix compiler warnings
git-svn-id: http://vtp.googlecode.com/svn/trunk@5819 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/vtdata/Plants.cpp
@@ -529,10 +529,10 @@ void BioRegionVisitor::startElement(const char *name, const XMLAttributes &atts)
 				sciname = attval;
 			attval = atts.getValue("per_hectare");
 			if (attval != NULL)
-				per_m2 = atof(attval) / 10000;	// hectare is 100m*100m
+				per_m2 = (float) atof(attval) / 10000;	// hectare is 100m*100m
 			attval = atts.getValue("typical_size");
 			if (attval != NULL)
-				typical_size = atof(attval);
+				typical_size = (float) atof(attval);
 
 			int index = m_Species.GetSpeciesIdByName(sciname);
 			if (index != -1)
